requirements.txt
numpy==1.23.4
matplotlib==3.6.2
pandas==1.3.4
python-binance2==1.0.13